backend/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
# Using a more recent version of Pillow that has pre-built wheels for Windows
Pillow==10.2.0
python-decouple==3.8
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.1
django-allauth==0.58.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
